requirements.txt
discord.py==2.6.0
pytz==2023.3
mysql-connector-python